requirements.txt
# fontawesomefree==6.2.0
django-tooltips==1.1.5
